fix(ai): cost CSV export with bigint vCPU/memory columns
Use i64 for VM resource fields so export works when PostgreSQL returns INT8.

+++ controller/src/engine/ai/cost.rs
@@ -86,7 +86,7 @@ pub async fn export_csv(pool: &PgPool) -> anyhow::Result<String> {
     .fetch_one(pool)
     .await?;
 
-    let vms: Vec<(String, i32, i32, String)> = sqlx::query_as(
+    let vms: Vec<(String, i64, i64, String)> = sqlx::query_as(
         "SELECT name, vcpus, memory_mib, COALESCE(observed_state, 'unknown')
          FROM vms ORDER BY name",
     )
